extract serialization
